Revert "Merge branch 'EliO/UpgradeTaproot' into 'develop'"
This reverts commit 483abb5c3541d2b98e577ad9462451c42f0beb2e, reversing
changes made to 08435ed124effcb46d0c8a477a4b2619326dec7f.

+++ aruw-mcb-project/src/aruwsrc/control/agitator/agitator_test_command.hpp
@@ -38,21 +38,18 @@ public:
 
     bool isReady() override { return true; };
 
-    void initialize() override
-    {
-        subsystem->setSetpoint(subsystem->getCurrentValue() + static_cast<float>(M_PI) / 2);
-    };
+    void initialize() override { subsystem->setSetpoint(subsystem->getCurrentValue() + M_PI / 2); };
 
-    void execute() override {};
+    void execute() override{};
 
-    void end(bool) override {};
+    void end(bool) override{};
 
     bool isFinished() const override
     {
         return tap::algorithms::compareFloatClose(
             this->subsystem->getSetpoint(),
             this->subsystem->getCurrentValue(),
-            static_cast<float>(M_PI) / 16);
+            M_PI / 16);
     };
 
     const char* getName() const override { return "agitator test command"; }
