+++ aruw-mcb-project/src/aruwsrc/robot/hero/hero_control.cpp
@@ -345,7 +345,7 @@ GovernorWithFallbackCommand<3> beybladeSlowWhenOutOfCombatCommand(
     true);
 IMUCalibrateDoneGovernor imuCalibrateDoneGovernor(drivers(), imuCalibrateCommand);
 
-user::TurretQuickTurnCommand turretUTurnCommand(&turret, static_cast<float>(M_PI));
+user::TurretQuickTurnCommand turretUTurnCommand(&turret, M_PI);
 
 GovernorLimitedCommand<1> turretUTurnCommandLimited(
     {&turret},
